requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.11.0
sentencepiece
huggingface-hub
pytorch-lightning>=2.0.0

openai-whisper

librosa
soundfile
webrtcvad
scipy
numpy
sounddevice

opencv-python
mediapipe
facenet-pytorch

TTS

python-dotenv
matplotlib
scikit-learn
gradio
